[Feat]: user properties 반환을 위한 response 객체 및 controller 구현
Related to: #538

+++ src/main/java/org/sopt/makers/internal/member/controller/dto/response/MemberPropertiesResponse.java
@@ -20,8 +20,8 @@ public record MemberPropertiesResponse(
 
         CoffeeChatStatus coffeeChatStatus,
 
-        Integer receivedCoffeeChatCount,
+        Long receivedCoffeeChatCount,
 
-        Integer sentCoffeeChatCount
+        Long sentCoffeeChatCount
 ) {
 }
